fix validation

+++ Services/Catalog/FreeCourse.Services.Catalog/Models/FeatureEntity.cs
@@ -2,6 +2,6 @@ namespace FreeCourse.Services.Catalog.Models
 {
     public class FeatureEntity
     {
-        public short Duration { get; set; }
+        public int Duration { get; set; }
     }
 }
